Split up controllers per Tab
I reversed the change of having a single document controller and re-created two different controllers, one for each tab. I think this makes more sense and prevents us from having to label FXML elements with T1 or T2 to distinguish which tab they belong to.
