+++ tabControllers/Tab2.fxml
@@ -7,10 +7,11 @@
 <?import javafx.scene.chart.PieChart?>
 <?import javafx.scene.control.Accordion?>
 <?import javafx.scene.control.Button?>
-<?import javafx.scene.control.CheckBox?>
 <?import javafx.scene.control.Menu?>
 <?import javafx.scene.control.MenuBar?>
 <?import javafx.scene.control.MenuItem?>
+<?import javafx.scene.control.ProgressIndicator?>
+<?import javafx.scene.control.Spinner?>
 <?import javafx.scene.control.SplitMenuButton?>
 <?import javafx.scene.control.SplitPane?>
 <?import javafx.scene.control.Tab?>
@@ -21,8 +22,10 @@
 <?import javafx.scene.control.ToolBar?>
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.web.WebView?>
+<?import javafx.scene.control.SpinnerValueFactory.IntegerSpinnerValueFactory?>
+<?import javafx.scene.control.* ?>
 
-<AnchorPane id="AnchorPane" prefHeight="647.0" prefWidth="787.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="coinTrack.FXMLDocumentController">
+<AnchorPane id="AnchorPane" prefHeight="647.0" prefWidth="787.0" xmlns="http://javafx.com/javafx/11.0.1" xmlns:fx="http://javafx.com/fxml/1" fx:controller="tabControllers.Tab2Controller">
     <children>
       <MenuBar prefHeight="28.0" prefWidth="787.0">
         <menus>
@@ -45,12 +48,6 @@
       </MenuBar>
       <ToolBar layoutX="14.0" layoutY="528.0" prefHeight="34.0" prefWidth="762.0">
         <items>
-            <SplitMenuButton fx:id="splitMenuT2" mnemonicParsing="false" text="Select">
-              <items>
-                <MenuItem fx:id="searchAllCoins" mnemonicParsing="false" onAction="#handleAllCoins" text="All Coins" />
-                <MenuItem fx:id="searchSingleCoins" mnemonicParsing="false" onAction="#handleSingleCoin" text="Single Coin" />
-              </items>
-            </SplitMenuButton>
             <SplitMenuButton fx:id="numCoins" mnemonicParsing="false" text="Amount">
               <items>
                 <MenuItem fx:id="numCoins5" mnemonicParsing="false" onAction="#handle5Coins" text="5" />
@@ -59,12 +56,12 @@
                   <MenuItem fx:id="numCoinsAll" mnemonicParsing="false" onAction="#handleTotalCoins" text="All" />
               </items>
             </SplitMenuButton>
+            <Spinner fx:id="spinnerT2" />
             <Button fx:id="scanBtnT2" accessibleHelp="Scan for updates on coin prices" defaultButton="true" mnemonicParsing="false" onAction="#handleScanT2" text="Scan" underline="true" />
-            <Button fx:id="clearBtnT2" onAction="#handleClearT2" mnemonicParsing="false" text="Clear" />
+            <Button fx:id="clearBtnT2" mnemonicParsing="false" onAction="#handleClearT2" text="Clear" />
             <TextField fx:id="searchFieldT2" />
           <Button fx:id="searchBtnT2" accessibleHelp="Search for a specific cryptocurrency" mnemonicParsing="false" onAction="#handleSearchT2" text="Search" />
-            <CheckBox fx:id="searchSymbolT2" mnemonicParsing="false" onAction="#handleSymbolSearch" text="Symbol" />
-            <CheckBox fx:id="searchNameT2" mnemonicParsing="false" onAction="#handleNameSearch" selected="true" text="Name" />
+            <ProgressIndicator fx:id="progIndT2" progress="0.0" />
         </items>
       </ToolBar>
       <SplitPane dividerPositions="0.8770161290322581" layoutY="28.0" orientation="VERTICAL" prefHeight="498.0" prefWidth="787.0">
@@ -98,9 +95,9 @@
                         </AnchorPane>
                       <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="160.0" prefWidth="100.0">
                            <children>
-                              <TabPane prefHeight="430.0" prefWidth="614.0" tabClosingPolicy="UNAVAILABLE">
+                              <TabPane fx:id="graphTabPane" prefHeight="430.0" prefWidth="614.0" tabClosingPolicy="UNAVAILABLE">
                                 <tabs>
-                                  <Tab text="Bar Chart">
+                                  <Tab fx:id="barChartTab" text="Bar Chart">
                                     <content>
                                       <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
                                              <children>
@@ -116,7 +113,7 @@
                                           </AnchorPane>
                                     </content>
                                   </Tab>
-                                  <Tab text="Pie Chart">
+                                  <Tab fx:id="pieChartTab" text="Pie Chart">
                                     <content>
                                       <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
                                              <children>
@@ -125,7 +122,7 @@
                                           </AnchorPane>
                                     </content>
                                   </Tab>
-                                    <Tab text="Bubble Chart">
+                                    <Tab fx:id="bubbleChartTab" text="Bubble Chart">
                                       <content>
                                         <AnchorPane minHeight="0.0" minWidth="0.0" prefHeight="180.0" prefWidth="200.0">
                                              <children>
